refactor: PR feedback - Move to new package, introduce method to get helper objects.
