+++ src/test/java/com/synopsys/integration/jenkins/wrapper/JenkinsProxyHelperTest.java
@@ -1,4 +1,4 @@
-package com.synopsys.integration.jenkins;
+package com.synopsys.integration.jenkins.wrapper;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
